moved iothub_mapper and controllerswitch to own package
