+++ kyt-api-server/iothubservice/iothub_mapper.go
@@ -1,4 +1,4 @@
-package openapi
+package iothubservice
 
 import (
 	"fmt"
